Bump deps and migrate conf generator to its own module
