+++ setup.cfg
@@ -60,31 +60,31 @@ console_scripts =
 
 [options.extras_require]
 docs =
-    furo>=2021.8.17b43
+    furo>=2022.1.2
     sphinx>=4.4
-    sphinx-argparse-cli>=1.7
+    sphinx-argparse-cli>=1.8.3
     sphinx-autodoc-typehints>=1.16
-    sphinx-copybutton>=0.4
-    sphinx-inline-tabs>=2021.4.11b9
-    sphinxcontrib-towncrier>=0.2.0a0
+    sphinx-copybutton>=0.5
+    sphinx-inline-tabs>=2022.1.2b11
+    sphinxcontrib-towncrier>=0.2.1a0
     towncrier>=21.3
 testing =
-    covdefaults>=1.2
+    covdefaults>=2.2
     devpi-client>=5.2
-    devpi-server>=6.1
-    distlib>=0.3.2
-    filelock>=3
+    devpi-server>=6.4
+    distlib>=0.3.4
+    filelock>=3.4
     flaky>=3.7
     freezegun>=1.1
-    psutil>=5.8
-    pytest>=6.2
-    pytest-cov>=2.12
-    pytest-mock>=3.6
-    pytest-xdist>=2.3
+    psutil>=5.9
+    pytest>=7
+    pytest-cov>=3
+    pytest-mock>=3.7
+    pytest-xdist>=2.5
     re-assert>=1.1
-    setuptools>=57
+    setuptools>=60
     setuptools-scm>=6
-    wheel>=0.36
+    wheel>=0.37
 
 [options.package_data]
 tox = py.typed
